Move dummy model to `api.py` and improve the /api/predict/ endpoint.
