+++ dockerfiles/api.dockerfile
@@ -10,7 +10,6 @@ WORKDIR /
 COPY deps/requirements_api.txt requirements.txt
 COPY pyproject.toml .
 COPY src/dtu_mlops_project/api.py .
-COPY src/dtu_mlops_project/model.py .
 
 RUN touch __init__.py
 RUN pip install -r requirements.txt --no-cache-dir
